client update for databricks on maggy
