+++ maggy/core/executors/tf_dist_executor.py
@@ -27,7 +27,7 @@ import tensorflow as tf
 from maggy import util, tensorboard
 from maggy.core.tf_patching.tf_modules import get_wrapped_model
 from maggy.experiment_config import TfDistributedConfig
-from maggy.core.rpc import Client
+from maggy.core.environment.client import Client
 from maggy.core.reporter import Reporter
 from maggy.core.environment.singleton import EnvSing
 
@@ -66,7 +66,13 @@ def dist_executor_fn(
 
         EnvSing.get_instance().set_ml_id(app_id, run_id)
         partition_id, _ = util.get_partition_attempt_id()
-        client = Client(server_addr, partition_id, 0, hb_interval, secret)
+        client = EnvSing.get_instance().get_client(
+            server_addr,
+            partition_id,
+            hb_interval,
+            secret,
+            socket.socket(socket.AF_INET, socket.SOCK_STREAM),
+        )
         log_file = log_dir + "/executor_" + str(partition_id) + ".log"
 
         reporter = Reporter(log_file, partition_id, 0, __builtin__.print)
